refactor(rust): move `resolve_id` to `rolldown_plugin`
