+++ crates/rolldown/src/types/mod.rs
@@ -9,5 +9,4 @@ pub mod linking_metadata;
 pub mod match_import_kind;
 pub mod module_render_output;
 pub mod namespace_alias;
-pub mod resolved_request_info;
 pub mod symbols;
