+++ crates/rolldown_common/src/types/mod.rs
@@ -22,6 +22,7 @@ pub mod rendered_chunk;
 pub mod rendered_module;
 pub mod resolved_export;
 pub mod resolved_path;
+pub mod resolved_request_info;
 pub mod stmt_info;
 pub mod symbol_ref;
 pub mod wrap_kind;
