+++ crates/rolldown_plugin/src/utils/mod.rs
@@ -1,0 +1,1 @@
+pub mod resolve_id_with_plugins;
